+++ Sourcery/Models/Class.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+// sourcery: skipDescription, skipEquatable
+class Class: Type {
+    override var kind: String { return "class" }
+}